requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.29.5
certifi==2023.7.22
charset-normalizer==3.3.1
dataclasses-json==0.6.1
Django==4.2.6
exceptiongroup==1.1.3
frozenlist==1.4.0
idna==3.4
isodate==0.6.1
jsonpatch==1.33
jsonpointer==2.4
langchain==0.0.326
langsmith==0.0.54
marshmallow==3.20.1
multidict==6.0.4
mypy-extensions==1.0.0
numpy==1.26.1
openai==0.28.1
packaging==23.2
Pillow==10.1.0
pydantic==2.4.2
pydantic_core==2.10.1
python-dotenv==1.0.0
pytube==15.0.0
PyYAML==6.0.1
requests==2.31.0
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.22
sqlparse==0.4.4
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
urllib3==2.0.7
whisper==1.1.10
yarl==1.9.2
